pyproject.toml
[project]
name = "chat"
version = "0.3.0"
description = "Default template for PDM package"
authors = [
    {name = "kimpass189", email = "kimpass189@gmail.com"},
]
dependencies = [
    "kafka-python>=2.0.2",
    "textual>=0.73.0",
    "requests>=2.32.3",
    "pandas>=2.0.3",
    "pyarrow>=17.0.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = true
